revert mistakes in scripts (#413)

+++ spruce/scripts/gameswitcher.sh
@@ -145,11 +145,17 @@ while : ; do
 
     # run switcher
     log_message "***** gameswitcher.sh: launching actual switcher executable" -v
-    cd $BIN_PATH
-    /mnt/SDCARD/spruce/bin64/gptokeyb -k "switcher" -c "./switcher.gptk" &
-    ./switcher "$IMAGES_FILE" "$GAMENAMES_FILE" $OPTIONS \
-    -dc "sed -i 'INDEXs/.*/removed/' $LIST_FILE"
-    kill -9 "$(pidof gptokeyb)"
+    if [ "$PLATFORM" = "A30" ]; then
+        cd $BIN_PATH
+        ./switcher "$IMAGES_FILE" "$GAMENAMES_FILE" $OPTIONS \
+        -dc "sed -i 'INDEXs/.*/removed/' $LIST_FILE"
+    else
+        cd $BIN_PATH
+        /mnt/SDCARD/spruce/bin64/gptokeyb -k "switcher" -c "./switcher.gptk" &
+        ./switcher "$IMAGES_FILE" "$GAMENAMES_FILE" $OPTIONS \
+        -dc "sed -i 'INDEXs/.*/removed/' $LIST_FILE"
+        kill -9 "$(pidof gptokeyb)"
+    fi
     # get return value
     RETURN_INDEX=$?
